requirements.txt
imbalanced-learn
joblib
matplotlib
mlflow==2.19.0
numpy
pandas
python-dotenv
scikit-learn
seaborn
